add libc++ build script update README with libc++ build instructions and samples fix mavericks SDK download link (pointed to 10.8) build xar only for <=10.5 add -g0 to the clang invocation command to avoid dsymutil from being run (debugging is not supported, but I guess you don't want to debug the resulting binaries anyway if you build them on a non OS X system) attempt to make the toolchain less path dependent (gcc still breaks though, because of hardcoded paths), but clang and cctools can be moved now update cctools to 845 add DISABLE_LTO_SUPPORT option (DISABLE_LTO_SUPPORT=1 ./build.sh) to disable linking against libLTO.so add support for 32 bit systems add FreeBSD support update PACKAGE script

+++ oclang/find_intrinsic_headers.sh
@@ -17,14 +17,17 @@ CLANG_DIR=`dirname \`which clang\``
 CLANG_INTRIN_DIR="$CLANG_DIR/../include/clang/$CLANG_VERSION/include"
 
 test ! -e "$CLANG_INTRIN_DIR" && CLANG_INTRIN_DIR="$CLANG_DIR/../lib/clang/$CLANG_VERSION/include"
+test ! -e "$CLANG_INTRIN_DIR" && CLANG_INTRIN_DIR="$CLANG_DIR/../include/clang/$CLANG_VERSION/include"
+test ! -e "$CLANG_INTRIN_DIR" && CLANG_INTRIN_DIR="$CLANG_DIR/../include/clang/$CLANG_VERSION"
 
 test -e "$CLANG_INTRIN_DIR" || { echo "can not find clang intrinsics directory" && exit 1; }
 test -f "$CLANG_INTRIN_DIR/xmmintrin.h" || { echo "xmmintrin.h does not exist in $CLANG_INTRIN_DIR" && exit 1; }
 
 echo "found clang intrinsic headers: $CLANG_INTRIN_DIR"
 
-ln -sf $CLANG_INTRIN_DIR/float.h $SDK_DIR/usr/include
-ln -sf $CLANG_INTRIN_DIR/stdarg.h $SDK_DIR/usr/include
+test -f $CLANG_INTRIN_DIR/float.h && ln -sf $CLANG_INTRIN_DIR/float.h $SDK_DIR/usr/include
+test -f $CLANG_INTRIN_DIR/stdarg.h && ln -sf $CLANG_INTRIN_DIR/stdarg.h $SDK_DIR/usr/include
+
 ln -sf $CLANG_INTRIN_DIR/*intrin*.h $SDK_DIR/usr/include
 ln -sf $CLANG_INTRIN_DIR/mm*.h $SDK_DIR/usr/include
 ln -sf $CLANG_INTRIN_DIR/*va*.h $SDK_DIR/usr/include
